+++ web/example/CheckBox.dart
@@ -56,6 +56,7 @@ grid.SlickGrid init(){
                   ..multiColumnSort= false
                   ..multiSelect=false
                   ..autoEdit=false
+                  ..enableColumnReorder = true
                   ..frozenColumn=2;
   grid.SlickGrid sg= new grid.SlickGrid.fromOpt(el,data,column,opt);
   RowSelectionModel rsm=new RowSelectionModel({'selectActiveRow':true});